requirements.txt
Django==5.2.11
gunicorn==23.0.0
whitenoise==6.9.0
Pillow==11.2.1
pillow-heif==0.21.0
requests==2.32.3
